re-allow ext-redis 6.1
